# openapi.yaml
openapi: 3.0.0
info:
  title: API de Charadas
  description: API para gerenciamento de charadas com autenticação via token
  version: 1.0.0
  contact:
    name: Joaquim

servers:
  - url: http://localhost:5000

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    Charada:
      type: object
      properties:
        id:
          type: integer
          example: 1
        pergunta:
          type: string
          example: "O que é, o que é: quanto mais tira, maior fica?"
        resposta:
          type: string
          example: "Buraco"

    LoginRequest:
      type: object
      required:
        - usuario
        - senha
      properties:
        usuario:
          type: string
          example: admin
        senha:
          type: string
          example: 123456

    LoginResponse:
      type: object
      properties:
        message:
          type: string
        token:
          type: string

    Message:
      type: object
      properties:
        message:
          type: string

    Error:
      type: object
      properties:
        error:
          type: string

paths:

  /:
    get:
      summary: Rota principal
      responses:
        '200':
          description: Informações da API
          content:
            application/json:
              schema:
                type: object
                properties:
                  api:
                    type: string
                  version:
                    type: string
                  autor:
                    type: string

  /login:
    post:
      summary: Realizar login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login realizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Não autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /charadas:
    get:
      summary: Listar todas as charadas
      responses:
        '200':
          description: Lista de charadas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Charada'

    post:
      summary: Criar nova charada
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - pergunta
                - resposta
              properties:
                pergunta:
                  type: string
                resposta:
                  type: string
      responses:
        '201':
          description: Criado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '400':
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /charadas/aleatoria:
    get:
      summary: Obter uma charada aleatória
      responses:
        '200':
          description: Charada aleatória
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Charada'

  /charadas/{id}:
    get:
      summary: Buscar charada por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Charada encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Charada'
        '404':
          description: Não encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      summary: Atualizar completamente uma charada
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Charada'
      responses:
        '200':
          description: Atualizado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '404':
          description: Não encontrada

    patch:
      summary: Atualizar parcialmente uma charada
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                pergunta:
                  type: string
                resposta:
                  type: string
      responses:
        '200':
          description: Atualizado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'

    delete:
      summary: Remover uma charada
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Removido com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '404':
          description: Não encontrada